Clean up test mods
